145 implement robot control functionality to interact with user

+++ SBR_SERV/serial_connector/src/type_message_esp32.rs
@@ -3,5 +3,5 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, Deserialize, Serialize)]
 pub struct MessageEsp32 {
     pub name: String,
-    pub data: i32,
+    pub data: f64,
 }
